[BACKLOG-19904] Refactoring AMD module id parsing logic into an util module.
